Replace AsyncJob with PendingJob
`AsyncJob` was more or less the same as the `concurrent.Future`,
which was actually unnecessary because...

* It initializes with `Future` object regardless of whether the job
goes to the running or the pending queue. But there's no need to
populate the `Future` when the job goes to the pending queue.
* The custom future-like object without special methods introduces
nothing good. It seems better for users to deal with the native
`concurrent.Future` object directly.

Therefore the replacing `PendingJob` class does not embrace `Future`,
and it is solely for storing a coroutine and its corresponding callback
in the pending queue.

Please see the reference #2.

This commit contains other tedious cleanups.